+++ src/common/util.ts
@@ -24,3 +24,17 @@ export async function prepareAuthorizationHeader(): Promise<string> {
     const loginResult = await login();
     return `Bearer ${loginResult && loginResult.accessToken}`;
 }
+
+export function randomStrings(length: number): string {
+    var result = '';
+    var characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    var charactersLength = characters.length;
+    for (var i = 0; i < length; i++) {
+        result += characters.charAt(Math.floor(Math.random() * charactersLength));
+    }
+    return result;
+}
+
+export function randomNumber(): number {
+    return Math.random() * 100;
+}