+++ src/extension.ts
@@ -1,8 +1,10 @@
 import * as vscode from 'vscode';
-import { connectToCentral } from './service/central';
+import { start, stop, test } from './service/central';
 import { login } from './service/login';
 
 export function activate(context: vscode.ExtensionContext) {
+
+	// start();
 
 	context.subscriptions.push(vscode.commands.registerCommand(
 		'iotc.start',
@@ -10,7 +12,7 @@ export function activate(context: vscode.ExtensionContext) {
 			const result = await vscode.window.showInformationMessage('Do you want to login to Azure and connect IoTCentral?', 'Yes', 'No');
 			if (result === 'Yes') {
 				await login();
-				await connectToCentral();
+				await start();
 				vscode.window.showInformationMessage('Connected to IoTCentral application https://vscode.azureiotcentral.com');
 			}
 			else {
@@ -21,10 +23,22 @@ export function activate(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(vscode.commands.registerCommand(
 		'iotc.stop',
-		() => {
+		async () => {
+			await stop();
 			vscode.window.showInformationMessage('Stopped');
+		}
+	));
+
+	context.subscriptions.push(vscode.commands.registerCommand(
+		'iotc.test',
+		async () => {
+			await login();
+			await test();
+			vscode.window.showInformationMessage('Test telemetry sent to IoTCentral');
 		}
 	));
 }
 
-export function deactivate() { }
+export function deactivate() {
+	// stop();
+}